refactor(distributed): own generic sdpa in cp sharder
