+++ tests/unit_tests/distributed/test_cp_utils_inputs_embeds.py
@@ -29,7 +29,7 @@ import contextlib
 import pytest
 import torch
 
-from nemo_automodel.components.distributed.context_parallel import transport as _ct
+from nemo_automodel.components.distributed.context_parallel import sharder as _cs
 from nemo_automodel.components.distributed.context_parallel.runtime import ContextParallelRuntime
 
 
@@ -76,8 +76,8 @@ def _prepare(device_mesh, batch, loss_mask=None):
 
 def test_xor_assertion_neither_present(monkeypatch):
     """Batch missing both input_ids AND inputs_embeds must raise AssertionError."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kw: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kw: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     batch = {"labels": torch.zeros(1, 4, dtype=torch.long)}  # neither present
     with pytest.raises(AssertionError, match="exactly one of"):
@@ -86,8 +86,8 @@ def test_xor_assertion_neither_present(monkeypatch):
 
 def test_xor_assertion_both_present(monkeypatch):
     """Batch with BOTH input_ids and inputs_embeds must raise AssertionError."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kw: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kw: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     batch = {
         "input_ids": torch.zeros(1, 4, dtype=torch.long),
@@ -106,8 +106,8 @@ def test_inputs_embeds_path_uses_embeds_as_primary_seq_tensor(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(1, 8, 16)  # [B=1, S=8, H=16]
@@ -129,8 +129,8 @@ def test_inputs_embeds_with_grad_is_sharded_out_of_place(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1, cp_rank=1)
     full_sequence = torch.arange(8.0, requires_grad=True)
@@ -160,8 +160,8 @@ def test_inputs_embeds_with_grad_is_sharded_out_of_place(monkeypatch):
 )
 def test_inputs_embeds_grad_sharding_uses_cp_submesh_rank_under_hsdp(monkeypatch, root_rank, cp_rank, expected):
     """HSDP root coordinates must not affect the rank used for CP sharding."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kw: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kw: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyHSDPDeviceMesh(root_rank=root_rank, cp_rank=cp_rank)
     full_sequence = torch.arange(8.0, requires_grad=True)
@@ -199,8 +199,8 @@ def test_inputs_embeds_with_grad_and_cp_padding_preserves_global_token_mean(monk
             captured.update(kwargs)
             return object()
 
-        monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-        monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+        monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+        monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
         embedding = torch.nn.Embedding(sequence_length, hidden_size, dtype=torch.float64)
         with torch.no_grad():
@@ -260,8 +260,8 @@ def test_input_ids_path_unchanged(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     input_ids = torch.zeros(1, 8, dtype=torch.long)
@@ -278,8 +278,8 @@ def test_position_ids_synthesized_from_inputs_embeds_seq_dim(monkeypatch):
     """When position_ids is missing AND inputs_embeds is the primary, the
     synthesized arange must use ``inputs_embeds.shape[1]`` (the seq dim of the
     embed tensor, not its hidden dim)."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kw: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kw: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(1, 12, 32)  # B=1, S=12, H=32
@@ -296,8 +296,8 @@ def test_position_ids_synthesized_from_inputs_embeds_seq_dim(monkeypatch):
 def test_position_ids_synthesized_for_each_batch_row(monkeypatch):
     """Synthesized 2D position_ids must match the batch dimension so later CP
     sharding keeps positions aligned with each sample."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kw: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kw: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(3, 8, 16)
@@ -319,8 +319,8 @@ def test_singleton_position_ids_expand_to_batch_size(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(2, 8, 16)
@@ -359,8 +359,8 @@ def test_inputs_embeds_path_preserves_padding_mask_in_cp_buffers(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(1, 8, 16)
@@ -387,8 +387,8 @@ def test_padding_pads_all_buffers_to_cp_divisor_multiple(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     seq_len = 6  # 6 % 4 = 2 -> pad to 8
@@ -425,8 +425,8 @@ def test_padding_labels_use_negative_100_int_buffers_use_zero(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     inputs_embeds = torch.ones(1, 6, 4)
@@ -463,8 +463,8 @@ def test_padding_handles_loss_mask_and_padding_mask(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     seq_len = 6
@@ -490,8 +490,8 @@ def test_padding_handles_loss_mask_and_padding_mask(monkeypatch):
 
 def test_additional_token_tensor_uses_prepared_layout(monkeypatch):
     """Teacher logits use the returned token layout instead of a dispatch argument."""
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", lambda **kwargs: object())
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", lambda **kwargs: object())
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     teacher_logits = torch.randn(1, 6, 5)
@@ -521,8 +521,8 @@ def test_padding_mask_pad_value_is_True_not_False(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     seq_len = 6  # 6 % 4 = 2 -> pad to 8
@@ -576,8 +576,8 @@ def test_padding_mirrors_padding_mask_back_into_batch(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     seq_len = 6  # 6 % 4 = 2 -> pad to 8
@@ -613,8 +613,8 @@ def test_padding_no_op_when_seq_already_aligned(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     seq_len = 8  # 8 % 4 == 0 -> no pad
@@ -638,8 +638,8 @@ def test_padding_input_ids_path_int_padding_with_zero(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)  # divisor = 4
     input_ids = torch.tensor([[1, 2, 3, 4, 5, 6]])
@@ -669,8 +669,8 @@ def test_inputs_embeds_3d_position_ids_seq_dim(monkeypatch):
         captured.update(kwargs)
         return object()
 
-    monkeypatch.setattr(_ct, "create_context_parallel_ctx", _fake_create_ctx)
-    monkeypatch.setattr(_ct, "get_train_context", lambda *a, **kw: "ctx")
+    monkeypatch.setattr(_cs, "_create_context_parallel_ctx", _fake_create_ctx)
+    monkeypatch.setattr(_cs, "_get_train_context", lambda *a, **kw: "ctx")
 
     device_mesh = _DummyDeviceMesh(cp_size=2, tp_size=1)
     inputs_embeds = torch.randn(1, 8, 16)
